Switch Postgres plugin to uuidv7 from ulid
Unfortunately, support for ULID is lacking in sqlx.

- https://github.com/pksunkara/pgx_ulid/issues/27
- https://github.com/dylanhart/ulid-rs/issues/69

We could convert ULIDs back and forth to UUIDs when touching them in
Rust, but I think that's too annoying. Instead, we can use the
better-supported UUID, specifically UUIDv7. To keep using short IDs for
URLs and such, we can encode them like
https://docs.rs/short-uuid/latest/short_uuid/.