Encode UUIDs in base58 in Valkey
Because we store the account ID as a string in Valkey, the encoding used
does affect the size in memory. Use base58 which takes us from 36
characters -> 32 characters without hyphens -> 22 characters used.

This savings will apply to any other UUIDs we deal with as well, and is
a much nicer presentation if a UUID ever needs to be shown to users.